move rune::attrs to rune::compile::attrs
